better handling of errors
Adding a member fails if they are already in another team, so we catch this, and remove them from the lists used to update groups.

Also downcase email addresses, as we had a users email address change case, and we tried to add them twice, resulting in a duplicate error.