Check for using both staff and student email addresses
Using both staff and student email addresses means one of these was manually created in the web interface.

It causes a conflict, as both cannot have the UPI as the external_ID, which breaks checking via the UPI.  A separated email lookup needed to be added to catch this conflict.