Code to process conf/exceptions.json
exceptions.json records manally maintained users.

It provides a record of Admin accounts

It allows adding of users that aren't in any research project

It allows overriding of the AD email field, if the IDP is responding with a different email address (e.g. happen when staff leave, or go back to being students).