Needed to make failded_to_add global
When we fail to add a team member, and later try to add them to a group, Dropbox throws an exception.

We were dealing with the case where the group add immediately followed the failure to add the member, bit not the case were they were also in another group that we tried to add them too much later.